+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Spells converted to new action system
  - All default spells changed
+ - Add Lockette integration
 
 ## 3.9.4